+++ src/SKIT.FlurlHttpClient.ByteDance.OceanEngine/Models/OAuth2/OAuth2AdvertiserGetResponse.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace SKIT.FlurlHttpClient.ByteDance.OceanEngine.Models
 {
     /// <summary>
-    /// <para>表示 [POST] /oauth2/advertiser/get 接口的响应。</para>
+    /// <para>表示 [GET] /oauth2/advertiser/get 接口的响应。</para>
     /// </summary>
     public class OAuth2AdvertiserGetResponse : OceanEngineResponse<OAuth2AdvertiserGetResponse.Types.Data>
     {
@@ -21,7 +21,7 @@ namespace SKIT.FlurlHttpClient.ByteDance.OceanEngine.Models
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("advertiser_id")]
                         [System.Text.Json.Serialization.JsonPropertyName("advertiser_id")]
-                        public int AdvertiserId { get; set; }
+                        public long AdvertiserId { get; set; }
 
                         /// <summary>
                         /// 获取或设置账号名称。
